+++ lib/src/configs.dart
@@ -1,0 +1,1 @@
+const kPport = 41762;