+++ CHANGELOG.md
@@ -7,3 +7,10 @@
 - Changed example file to `main.dart` so the *Example* tab shows up
 
 - Removed `Getting Started` section from *README*
+
+## 1.0.0
+
+### Breaking change
+
+- Re-written package from scratch. 
+- New package, new usage, new example